// src/types/foodcheck.ts
/**
 * Foodcheck type definitions
 */

export interface Foodcheck {
    guild_upgrade_id: number;   // ID in Guild Bank
    name: string;               // Pretty Name
    icon: string;               // Icon URL
    wiki_url: string;           // Wiki URL
    gw2_efficiency_url: string; // GW2 Efficiency Crafting URL
}